# Split codec from scheme