--- a/cmd/libs/go2idl/client-gen/testdata/apis/testgroup/v1/register.go
+++ b/cmd/libs/go2idl/client-gen/testdata/apis/testgroup/v1/register.go
@@ -17,15 +17,12 @@ limitations under the License.
 package v1
 
 import (
-	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/api/unversioned"
 	"k8s.io/kubernetes/pkg/api/v1"
 	"k8s.io/kubernetes/pkg/runtime"
 )
 
 var SchemeGroupVersion = unversioned.GroupVersion{Group: "testgroup", Version: "v1"}
-
-var Codec = runtime.CodecFor(api.Scheme, SchemeGroupVersion)
 
 func AddToScheme(scheme *runtime.Scheme) {
 	// Add the API to Scheme.