--- a/contrib/mesos/pkg/podutil/gzip.go
+++ b/contrib/mesos/pkg/podutil/gzip.go
@@ -25,6 +25,7 @@ import (
 	"k8s.io/kubernetes/pkg/api"
 	_ "k8s.io/kubernetes/pkg/api/install"
 	"k8s.io/kubernetes/pkg/api/v1"
+	"k8s.io/kubernetes/pkg/runtime"
 )
 
 func Gzip(pods <-chan *api.Pod) ([]byte, error) {
@@ -32,7 +33,7 @@ func Gzip(pods <-chan *api.Pod) ([]byte, error) {
 }
 
 func gzipList(list *api.PodList) ([]byte, error) {
-	raw, err := v1.Codec.Encode(list)
+	raw, err := runtime.Encode(api.Codecs.LegacyCodec(v1.SchemeGroupVersion), list)
 	if err != nil {
 		return nil, err
 	}
@@ -68,7 +69,7 @@ func gunzipList(gzipped []byte) (*api.PodList, error) {
 		return nil, err
 	}
 
-	obj, err := api.Scheme.Decode(raw)
+	obj, err := runtime.Decode(api.Codecs.UniversalDecoder(), raw)
 	if err != nil {
 		return nil, err
 	}