--- a/contrib/mesos/pkg/podutil/io.go
+++ b/contrib/mesos/pkg/podutil/io.go
@@ -27,6 +27,7 @@ import (
 	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/api/v1"
 	"k8s.io/kubernetes/pkg/api/validation"
+	"k8s.io/kubernetes/pkg/runtime"
 	utilyaml "k8s.io/kubernetes/pkg/util/yaml"
 )
 
@@ -41,7 +42,7 @@ func WriteToDir(pods <-chan *api.Pod, destDir string) error {
 			log.Warningf("skipping static pod %s/%s that had no filename", p.Namespace, p.Name)
 			continue
 		}
-		raw, err := v1.Codec.Encode(p)
+		raw, err := runtime.Encode(api.Codecs.LegacyCodec(v1.SchemeGroupVersion), p)
 		if err != nil {
 			log.Errorf("failed to encode static pod as v1 object: %v", err)
 			continue
@@ -105,7 +106,7 @@ func tryDecodeSinglePod(data []byte) (parsed bool, pod *api.Pod, err error) {
 	if err != nil {
 		return false, nil, err
 	}
-	obj, err := api.Scheme.Decode(json)
+	obj, err := runtime.Decode(api.Codecs.UniversalDecoder(), json)
 	if err != nil {
 		return false, pod, err
 	}