--- a/hack/lib/util.sh
+++ b/hack/lib/util.sh
@@ -282,7 +282,7 @@ kube::util::group-version-to-pkg-path() {
   # moving the results to pkg/apis/api.
   case "${group_version}" in
     # both group and version are "", this occurs when we generate deep copies for internal objects of the legacy v1 API.
-    /)
+    __internal)
       echo "api"
       ;;
     v1)
@@ -292,7 +292,7 @@ kube::util::group-version-to-pkg-path() {
       echo "api/unversioned"
       ;;
     *)
-      echo "apis/${group_version}"
+      echo "apis/${group_version%__internal}"
       ;;
   esac
 }