--- a/pkg/api/meta/restmapper.go
+++ b/pkg/api/meta/restmapper.go
@@ -23,6 +23,7 @@ import (
 	"strings"
 
 	"k8s.io/kubernetes/pkg/api/unversioned"
+	"k8s.io/kubernetes/pkg/runtime"
 	"k8s.io/kubernetes/pkg/util/sets"
 )
 
@@ -84,7 +85,7 @@ type DefaultRESTMapper struct {
 
 var _ RESTMapper = &DefaultRESTMapper{}
 
-// VersionInterfacesFunc returns the appropriate codec, typer, and metadata accessor for a
+// VersionInterfacesFunc returns the appropriate typer, and metadata accessor for a
 // given api version, or an error if no such api version exists.
 type VersionInterfacesFunc func(version unversioned.GroupVersion) (*VersionInterfaces, error)
 
@@ -92,7 +93,7 @@ type VersionInterfacesFunc func(version unversioned.GroupVersion) (*VersionInter
 // to a resource name and back based on the objects in a runtime.Scheme
 // and the Kubernetes API conventions. Takes a group name, a priority list of the versions
 // to search when an object has no default version (set empty to return an error),
-// and a function that retrieves the correct codec and metadata for a given version.
+// and a function that retrieves the correct metadata for a given version.
 func NewDefaultRESTMapper(defaultGroupVersions []unversioned.GroupVersion, f VersionInterfacesFunc) *DefaultRESTMapper {
 	resourceToKind := make(map[unversioned.GroupVersionResource]unversioned.GroupVersionKind)
 	kindToPluralResource := make(map[unversioned.GroupVersionKind]unversioned.GroupVersionResource)
@@ -251,6 +252,9 @@ func (m *DefaultRESTMapper) ResourceFor(resource unversioned.GroupVersionResourc
 
 func (m *DefaultRESTMapper) KindsFor(input unversioned.GroupVersionResource) ([]unversioned.GroupVersionKind, error) {
 	resource := input.GroupVersion().WithResource(strings.ToLower(input.Resource))
+	if resource.Version == runtime.APIVersionInternal {
+		resource.Version = ""
+	}
 
 	hasResource := len(resource.Resource) > 0
 	hasGroup := len(resource.Group) > 0
@@ -412,7 +416,7 @@ func (m *DefaultRESTMapper) RESTMapping(gk unversioned.GroupKind, versions ...st
 	var gvk *unversioned.GroupVersionKind
 	hadVersion := false
 	for _, version := range versions {
-		if len(version) == 0 {
+		if len(version) == 0 || version == runtime.APIVersionInternal {
 			continue
 		}
 
@@ -472,7 +476,6 @@ func (m *DefaultRESTMapper) RESTMapping(gk unversioned.GroupKind, versions ...st
 		GroupVersionKind: *gvk,
 		Scope:            scope,
 
-		Codec:            interfaces.Codec,
 		ObjectConvertor:  interfaces.ObjectConvertor,
 		MetadataAccessor: interfaces.MetadataAccessor,
 	}