--- a/pkg/api/testing/compat/compatibility_tester.go
+++ b/pkg/api/testing/compat/compatibility_tester.go
@@ -26,7 +26,6 @@ import (
 	"strings"
 	"testing"
 
-	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/api/unversioned"
 	"k8s.io/kubernetes/pkg/kubectl"
 	"k8s.io/kubernetes/pkg/runtime"
@@ -48,7 +47,7 @@ func TestCompatibility(
 ) {
 
 	// Decode
-	codec := runtime.CodecFor(api.Scheme, version)
+	codec := api.Codecs.LegacyCodec(version)
 	obj, err := runtime.Decode(codec, input)
 	if err != nil {
 		t.Fatalf("Unexpected error: %v", err)