--- a/pkg/api/testing/compat/compatibility_tester.go
+++ b/pkg/api/testing/compat/compatibility_tester.go
@@ -26,6 +26,7 @@ import (
 	"strings"
 	"testing"
 
+	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/api/unversioned"
 	"k8s.io/kubernetes/pkg/kubectl"
 	"k8s.io/kubernetes/pkg/runtime"