--- a/pkg/api/v1/register.go
+++ b/pkg/api/v1/register.go
@@ -17,7 +17,6 @@ limitations under the License.
 package v1
 
 import (
-	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/api/unversioned"
 	"k8s.io/kubernetes/pkg/runtime"
 )
@@ -27,9 +26,6 @@ const GroupName = ""
 
 // SchemeGroupVersion is group version used to register these objects
 var SchemeGroupVersion = unversioned.GroupVersion{Group: GroupName, Version: "v1"}
-
-// Codec encodes internal objects to the v1 scheme
-var Codec = runtime.CodecFor(api.Scheme, SchemeGroupVersion)
 
 func AddToScheme(scheme *runtime.Scheme) {
 	// Add the API to Scheme.