--- a/pkg/apis/abac/types.go
+++ b/pkg/apis/abac/types.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package api
+package abac
 
 import "k8s.io/kubernetes/pkg/api/unversioned"