… api/swagger-spec/api.json
… api/swagger-spec/apis.json
… api/swagger-spec/extensions.json
… api/swagger-spec/v1.json
… api/swagger-spec/v1beta1.json
… cmd/kube-apiserver/app/server_test.go
… cmd/libs/go2idl/client-gen/testoutput/testgroup/unversioned/testgroup_test.go
… contrib/mesos/pkg/executor/executor_test.go
… docs/api-reference/extensions/v1beta1/operations.html
… docs/api-reference/v1/operations.html
… examples/examples_test.go
… examples/https-nginx/make_secret.go
… examples/sharing-clusters/make_secret.go
… hack/after-build/update-generated-deep-copies.sh
… pkg/api/conversion_test.go
… pkg/api/copy_test.go
… pkg/api/deep_copy_test.go
… pkg/api/install/install_test.go
… pkg/api/meta/meta_test.go
… pkg/api/meta/restmapper_test.go
… pkg/api/serialization_test.go
… pkg/api/testapi/testapi_test.go
… pkg/api/v1/conversion_generated.go
… pkg/api/v1/defaults_test.go
… pkg/api/validation/schema_test.go
… pkg/apimachinery/registered/registered_test.go
… pkg/apis/abac/register.go
… pkg/apis/abac/v0/conversion_test.go
… pkg/apis/abac/v0/register.go
… pkg/apis/abac/v1beta1/register.go