Move all serialization (codec) responsibilities out of the scheme. Separate serialization
responsibilities of codec from versioning responsibilities of codec.
Enforce that all objects have a version (add __internal and
__unversioned to represent those).  Remove runtime.EmbeddedObject in favor
of pure runtime.Object -> runtime.RawExtension.

Disable the automatic encoding of runtime.Object during conversion - instead, require well behaved clients to interpret []runtime.Object themselves.  This removes the need for conversion to be aware of codecs (and of objects that the conversion may not have registered).

Dramatically reduce the signature for Codec. On Decode, pass a GVK tuple that
represents the desired decoding. Return from the decode the actual origin type
that the user presented (will allow resthandler to do checking orthogonal to
behavior in codec). Serializer and Codec look the same - the Codec type is 
for cases where conversion is in play, the serializer is where only raw wire
format is considered. Decode now returns enough info to make the right 
decisions at both levels

Begins addressing #7111 and required to support protobuf #17854 (cannot
assume only one serialization type and 1 codec per version)