@deads2k 2015-11-30
Any chance of  breaking the codec method signature changes in a separate pull?