@smarterclayton 2015-11-30
It's far too invasive - I want to get a fully working system back before I
do that.  However, go ahead and make your changes, assuming that I will
pick up purely mechanical string -> GVK changes.  Identify any places where
you think the assumptions are changing.

On Mon, Nov 30, 2015 at 7:48 AM, David Eads <notifications@github.com>
wrote:

> Any chance of breaking the codec method signature changes in a separate
> pull?
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/17922#issuecomment-160619490>
> .
>