@lavalamp 2015-12-02
I'm sure I missed things, hopefully not big things.

The nuance of Codec and Serializer was not at all obvious to me, but I can't think of better names off the top of my head.

Is Scheme still doing any useful work?

I think the stuff in the latest package needs to be reflected in the testapi package, or extracted to some common place (I think that sort of thing was the intended purpose of Scheme?) for use by both.