@smarterclayton 2015-12-02
Agree something needs to be done with latest and testapi.  As I'm walking
the changes out to packages I'm noticing the spots to change, will make
sure those get highlighted.


On Dec 1, 2015, at 7:49 PM, Daniel Smith <notifications@github.com> wrote:

I'm sure I missed things, hopefully not big things.

The nuance of Codec and Serializer was not at all obvious to me, but I
can't think of better names off the top of my head.

Is Scheme still doing any useful work?

I think the stuff in the latest package needs to be reflected in the
testapi package, or extracted to some common place (I think that sort of
thing was the intended purpose of Scheme?) for use by both.

—
Reply to this email directly or view it on GitHub
<https://github.com/kubernetes/kubernetes/pull/17922#issuecomment-161144016>
.