@smarterclayton 2015-12-06
@deads2k as part of the further refactoring I made unversioned.GroupVersion tolerate 1 segment and treat it as version, whereas two segments is always group/version.  Not every scheme we have is going to have its own groups or need them.  Arbitrary kube like objects with versions we don't recognize are still valid.  So "bugaboo" -> unversioned.GroupVersion{Version: "bugaboo"} -> "bugaboo".  Internal versions are fixed now because they are no longer empty string, and unversioned as we discussed are still in their own distinct version (right now, unversioned.GroupVersion{Version: "v1"}.

Other changes:

* As a follow on to this I'm going to remove the type registry bits of conversion/scheme and move them fully into runtime/scheme.  Conversion is about arbitrary interface{} -> interface{}, runtime/scheme is about runtime.Object, GVK, and conversions between logical types.
* TypeMeta being cleared internally -> externally is now handled by conversion
* Unversioned types are now registered specially into the scheme - during ConvertToVersion they are treated as a no-op.
* Created "LegacyCodec" with the same semantics as the existing codecs to ease the transition - this method is not used in the server and will be removed from use in clients (since both need to negotiate).