@smarterclayton 2015-12-06
@deads2k also note I removed "internalGroupVersions" from scheme - now that all internal have __internal, there is no ambiguity on empty string