@deads2k 2015-12-07
> @deads2k as part of the further refactoring I made unversioned.GroupVersion tolerate 1 segment and treat it as version, whereas two segments is always group/version. Not every scheme we have is going to have its own groups or need them. Arbitrary kube like objects with versions we don't recognize are still valid. So "bugaboo" -> unversioned.GroupVersion{Version: "bugaboo"} -> "bugaboo".

I don't think I like this.  I think objects should be forced to have a non-empty group or they should be considered invalid.  Also, if we were going to allow it, I would claim that "/the-version" is the proper single-string expression of a version without a group.  That eliminates any ambiguity.  They're doing a foolish thing, I don't feel sorry about making them specify a foolish looking value.