@smarterclayton 2015-12-07
Group is optional, and it has been.  You don't need groups if you aren't
part of the core Kube body.  If you have groups, great.  But empty group is
itself a group - it's the group of resources that don't care about other
people's namespace.  We can't force a unified namespace for all use cases.

On Mon, Dec 7, 2015 at 7:40 AM, David Eads <notifications@github.com> wrote:

> @deads2k <https://github.com/deads2k> as part of the further refactoring
> I made unversioned.GroupVersion tolerate 1 segment and treat it as version,
> whereas two segments is always group/version. Not every scheme we have is
> going to have its own groups or need them. Arbitrary kube like objects with
> versions we don't recognize are still valid. So "bugaboo" ->
> unversioned.GroupVersion{Version: "bugaboo"} -> "bugaboo".
>
> I don't think I like this. I think objects should be forced to have a
> non-empty group or they should be considered invalid. Also, if we were
> going to allow it, I would claim that "/the-version" is the proper
> single-string expression of a version without a group. That eliminates any
> ambiguity. They're doing a foolish thing, I don't feel sorry about making
> them specify a foolish looking value.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/17922#issuecomment-162513346>
> .
>