@deads2k 2015-12-07
> Group is optional, and it has been.  You don't need groups if you aren't
> part of the core Kube body.  If you have groups, great.  But empty group is
> itself a group - it's the group of resources that don't care about other
> people's namespace.  We can't force a unified namespace for all use cases.

This conflicts with @lavalamp's assertion that the only set of kinds that will ever have `group=""` is the legacy kube types.