@smarterclayton 2015-12-07
In kube.  But probably not in generic consumption of this library by other
codebases.  As a compromise (discussed offline) I'm going to have "group
defaulting per kind" which given a kind resolves it to a group - all
objects in the Kube group and the OpenShift groups will be required to have
this.  The serializer will convert those out.

On Mon, Dec 7, 2015 at 10:19 AM, David Eads <notifications@github.com>
wrote:

> Group is optional, and it has been. You don't need groups if you aren't
> part of the core Kube body. If you have groups, great. But empty group is
> itself a group - it's the group of resources that don't care about other
> people's namespace. We can't force a unified namespace for all use cases.
>
> This conflicts with @lavalamp <https://github.com/lavalamp>'s assertion
> that the only set of kinds that will ever have group="" is the legacy
> kube types.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/17922#issuecomment-162554883>
> .
>