@smarterclayton 2015-12-20
Some benchmark numbers before and after - the simpler Decode and Conversion paths show significant improvement:

```
benchmark                                      old ns/op     new ns/op     delta
BenchmarkPodConversion-8                       32078         23372         -27.14%
BenchmarkNodeConversion-8                      35751         28459         -20.40%
BenchmarkReplicationControllerConversion-8     26166         20244         -22.63%
BenchmarkEncodeCodec-8                         29595         28142         -4.91%
BenchmarkDecodeCodec-8                         88175         85287         -3.28%
BenchmarkDecodeIntoCodec-8                     219044        70571         -67.78%

benchmark                                      old allocs     new allocs     delta
BenchmarkPodConversion-8                       161            125            -22.36%
BenchmarkNodeConversion-8                      196            160            -18.37%
BenchmarkReplicationControllerConversion-8     139            103            -25.90%
BenchmarkEncodeCodec-8                         56             59             +5.36%
BenchmarkDecodeCodec-8                         347            336            -3.17%
BenchmarkDecodeIntoCodec-8                     1039           259            -75.07%

benchmark                                      old bytes     new bytes     delta
BenchmarkPodConversion-8                       12757         11637         -8.78%
BenchmarkNodeConversion-8                      13126         12005         -8.54%
BenchmarkReplicationControllerConversion-8     11829         10708         -9.48%
BenchmarkEncodeCodec-8                         5223          5726          +9.63%
BenchmarkDecodeCodec-8                         17679         19504         +10.32%
BenchmarkDecodeIntoCodec-8                     32991         13136         -60.18%
```